requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.42.2
Django==4.2.16
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==23.0.0
inflection==0.5.1
kombu==5.4.2
mysql-connector-python==9.2.0
mysqlclient==2.2.6
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0